Furniture placement much better; exits add blocks which are ignored by furniture and vice versa. In fact, the internal exit list may now be redundant, since exits are added as blocked areas.
Stairs, however, still are problematic - they should be blocked by
furniture as should the exits they require. An alternative would be
placing the exits before furnishing the rooms, but this is not viable
because of how exits are done - any room could gain an exit later on.

Buttresses should not overwrite rooms, except for gardens?